LPD-27612 Improve object actions pages and methods
