+++ modules/test/playwright/pages/object-web/EditObjectDefinitionPage.ts
@@ -6,16 +6,10 @@
 import {Locator, Page} from '@playwright/test';
 
 export class EditObjectDefinitionPage {
-	readonly actionsTab: Locator;
 	readonly layoutsTab: Locator;
 
 	constructor(page: Page) {
-		this.actionsTab = page.getByRole('link', {name: 'Actions'});
 		this.layoutsTab = page.getByRole('link', {name: 'Layouts'});
-	}
-
-	async openActionsTab() {
-		await this.actionsTab.click();
 	}
 
 	async openLayoutsTab() {
