+++ modules/test/playwright/pages/object-web/object-action/ViewObjectActionsPage.ts
@@ -5,11 +5,27 @@
 
 import {Locator, Page} from '@playwright/test';
 
+import {ViewObjectDefinitionsPage} from '../ViewObjectDefinitionsPage';
+
 export class ViewObjectActionsPage {
+	readonly actionsTabItem: Locator;
 	readonly addObjectActionButton: Locator;
+	readonly viewObjectDefinitionsPage: ViewObjectDefinitionsPage;
 
 	constructor(page: Page) {
+		this.actionsTabItem = page.getByRole('link', {name: 'Actions'});
 		this.addObjectActionButton = page.getByLabel('Add Object Action');
+		this.viewObjectDefinitionsPage = new ViewObjectDefinitionsPage(page);
+	}
+
+	async goto(objectDefinitionLabel: string) {
+		await this.viewObjectDefinitionsPage.goto();
+
+		await this.viewObjectDefinitionsPage.clickEditObjectDefinitionLink(
+			objectDefinitionLabel
+		);
+
+		await this.actionsTabItem.click();
 	}
 
 	async openObjectActionSidePanel() {
